fix end screen breaking question timing
basically what was happening is that even when exiting from the end screen, the previous timing thread for the next quesiton was still running, and so if you entered again say 4 seconds later, 6 seconds after you entered the question the timing thread will return and display the answer.
what this fix does is not restart the timing thread for another question when the question received is the end screen.